Fixed mktemp invocation to work on MacOS.
svn path=/branches/tp2/; revision=20890

+++ installer/src/main/resources/wrappers/wrapper.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 
-tmp=`mktemp -d`
+tmp=`mktemp -d -t install.XXXX`
+if [ $? -ne 0 ]; then
+    echo "Can't create temporary directory, exiting..."
+    exit 1
+fi
 echo "Extracting archive, please wait..."
-tail +15l $0 > /$tmp/tmp.jar
+tail +19l $0 > $tmp/tmp.jar
 cd $tmp
 $JAVA_HOME/bin/jar xvf tmp.jar 
 rm tmp.jar
@@ -10,5 +14,5 @@ chmod ugo+x product-installer.sh
 chmod ugo+x install/bin/engine-wrapper
 echo "InstallHome.directory.INSTALL_HOME=$HOME/glassfish-v3tp2" > install.properties
 sh product-installer.sh
-rm -rf * 
+rm -rf $tmp/* 
 exit $?
